+++ src/components/Form/index.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { AppContext } from '../../AppContext';
+import { AppContext } from '../../AppContextReact';
 import Step from '../Step';
 import Breadcrumb from '../Breadcrumb';
 import { validateField, validateStep } from './validation';
